hotfix: 책 삭제시 퀴즈도 같이 삭제하도록 변경
