+++ src/main/java/com/bookbla/americano/domain/memberask/service/impl/MemberAskServiceImpl.java
@@ -1,0 +1,53 @@
+package com.bookbla.americano.domain.memberask.service.impl;
+
+import com.bookbla.americano.base.exception.BaseException;
+import com.bookbla.americano.domain.member.repository.entity.Member;
+import com.bookbla.americano.domain.member.repository.MemberRepository;
+import com.bookbla.americano.domain.memberask.controller.dto.request.MemberAskCreateRequest;
+import com.bookbla.americano.domain.memberask.controller.dto.request.MemberAskUpdateRequest;
+import com.bookbla.americano.domain.memberask.controller.dto.response.MemberAskResponse;
+import com.bookbla.americano.domain.memberask.exception.MemberAskExceptionType;
+import com.bookbla.americano.domain.memberask.repository.MemberAskRepository;
+import com.bookbla.americano.domain.memberask.repository.entity.MemberAsk;
+import com.bookbla.americano.domain.memberask.service.MemberAskService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class MemberAskServiceImpl implements MemberAskService {
+
+    private final MemberAskRepository memberAskRepository;
+    private final MemberRepository memberRepository;
+
+    @Override
+    public MemberAskResponse createMemberAsk(Long memberId, MemberAskCreateRequest memberAskCreateRequest) {
+        Member member = memberRepository.getByIdOrThrow(memberId);
+        MemberAsk memberAsk = memberAskCreateRequest.toMemberAskWith(member);
+
+        MemberAsk savedMemberAsk = memberAskRepository.save(memberAsk);
+
+        return MemberAskResponse.from(savedMemberAsk);
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public MemberAskResponse readMemberAsk(Long memberId) {
+        Member member = memberRepository.getByIdOrThrow(memberId);
+        MemberAsk memberAsk  = memberAskRepository.findByMember(member)
+                .orElseThrow(() -> new BaseException(MemberAskExceptionType.NOT_REGISTERED_MEMBER));
+
+        return MemberAskResponse.from(memberAsk);
+    }
+
+    @Override
+    public void updateMemberAsk(Long memberId, MemberAskUpdateRequest memberAskUpdateRequest) {
+        Member member = memberRepository.getByIdOrThrow(memberId);
+        MemberAsk memberAsk  = memberAskRepository.findByMember(member)
+                .orElseThrow(() -> new BaseException(MemberAskExceptionType.NOT_REGISTERED_MEMBER));
+
+        memberAsk.updateContent(memberAskUpdateRequest.getContents());
+    }
+}
